requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
clr==1.0.3
comtypes==1.1.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.6
django-mysql==2.1.1
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
django-sslserver==0.20
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
mysqlclient==1.3.12
openapi-codec==1.3.2
psutil==5.4.1
psycopg2==2.7.3.2
Pygments==2.2.0
PyJWT==1.5.3
pypiwin32==220
pytz==2017.3
PyYAML==3.12
requests==2.18.4
simplejson==3.12.0
six==1.11.0
tqdm==4.19.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
